+++ owncloudApp/src/main/java/com/owncloud/android/presentation/ui/settings/fragments/SettingsSecurityFragment.kt
@@ -129,8 +129,11 @@ class SettingsSecurityFragment : PreferenceFragmentCompat() {
         prefAccessDocumentProvider = findPreference(PREFERENCE_ACCESS_FROM_DOCUMENT_PROVIDER)
         prefTouchesWithOtherVisibleWindows = findPreference(PREFERENCE_TOUCHES_WITH_OTHER_VISIBLE_WINDOWS)
 
+        prefPasscode?.isVisible = !securityViewModel.isSecurityEnforcedEnabled()
+        prefPattern?.isVisible = !securityViewModel.isSecurityEnforcedEnabled()
+
         // Passcode lock
-        prefPasscode?.setOnPreferenceChangeListener { preference: Preference?, newValue: Any ->
+        prefPasscode?.setOnPreferenceChangeListener { _: Preference?, newValue: Any ->
             if (securityViewModel.isPatternSet()) {
                 showMessageInSnackbar(getString(R.string.pattern_already_set))
             } else {
@@ -266,6 +269,7 @@ class SettingsSecurityFragment : PreferenceFragmentCompat() {
         private const val SCREEN_SECURITY = "security_screen"
         const val PREFERENCE_ACCESS_FROM_DOCUMENT_PROVIDER = "access_from_document_provider"
         const val PREFERENCE_TOUCHES_WITH_OTHER_VISIBLE_WINDOWS = "touches_with_other_visible_windows"
+        const val EXTRAS_LOCK_ENFORCED = "EXTRAS_LOCK_ENFORCED"
     }
 
 }