+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Summary
 * Bugfix - Small glitch when side menu is full of accounts: [#3437](https://github.com/owncloud/android/pull/3437)
 * Enhancement - Lock delay for app: [#3344](https://github.com/owncloud/android/issues/3344)
 * Enhancement - Allow access from document provider preference: [#3379](https://github.com/owncloud/android/issues/3379)
+* Enhancement - Security enforced: [#3434](https://github.com/owncloud/android/pull/3434)
 
 Details
 -------
@@ -41,6 +42,14 @@ Details
 
    https://github.com/owncloud/android/issues/3379
    https://github.com/owncloud/android/pull/3384
+
+* Enhancement - Security enforced: [#3434](https://github.com/owncloud/android/pull/3434)
+
+   A new branding/MDM option has been added to make app lock via passcode or pattern compulsory,
+   whichever the user chooses.
+
+   https://github.com/owncloud/android/issues/3400
+   https://github.com/owncloud/android/pull/3434
 
 Changelog for ownCloud Android Client [2.19.0] (2021-11-15)
 =======================================